+++ Common/DebuffOverhaul/Buffs/Poisoned.cs
@@ -1,14 +1,13 @@
-﻿using SpiritReforged.DebuffOverhaul.Common;
-using SpiritReforged.Common.Particle;
-using SpiritReforged.DebuffOverhaul.Content.Particles;
+﻿using SpiritReforged.Common.Particle;
+using SpiritReforged.Content.Particles.Basic;
 
-namespace SpiritReforged.DebuffOverhaul.Content.Buffs;
+namespace SpiritReforged.Common.DebuffOverhaul.Buffs;
 
-public class Bleeding : DoTExtension
+public class Poisoned : DoTExtension
 {
-    public override Settings LocalSettings => new(1f, 800);
+    public override Settings LocalSettings => new(0.1f, 500);
 
-    public override void Load() => Handler.Register(this, BuffID.Bleeding);
+    public override void Load() => Handler.Register(this, BuffID.Poisoned);
 
     public override void PostDrawHealthBar(SpriteBatch spriteBatch, HealthBarHook.Options options)
     {
@@ -17,7 +16,7 @@ public class Bleeding : DoTExtension
         float fadeout = MathHelper.Min(BuffTime / 30f, 1);
         float lightness = options.Lightness * 2;
         Rectangle bounds = new(0, 0, (int)(front.Width * progress), front.Height);
-        Color color = Color.Red;
+        Color color = new(0.6f, 1f, 0.2f);
 
         HealthBarHook.DrawSimpleBar(spriteBatch, front, options.Position, bounds, options.Scale, color * fadeout * lightness);
 
@@ -27,12 +26,12 @@ public class Bleeding : DoTExtension
 
         spriteBatch.Draw(bubble, endPosition, source, color * lightness, 0, source.Size() / 2, options.Scale, default, 0);
 
-        if ((int)Main.timeForVisualEffects % 25 == 0 && fadeout == 1)
-            TerrariaParticles.OverHealthBars.Add(new DropParticle(40, color * lightness, NPC)
-            {
-                LocalPosition = endPosition + Main.screenPosition - NPC.Center,
-                Scale = new Vector2(0.7f) * options.Scale,
-                AccelerationPerFrame = new(0, 0.02f)
-            });
+        if ((int)Main.timeForVisualEffects % 18 == 0 && fadeout == 1)
+			TerrariaParticles.OverHealthBars.Add(new BubbleParticle(40, color * lightness, NPC)
+			{
+				LocalPosition = endPosition + Main.screenPosition - NPC.Center,
+				Scale = new Vector2(0.8f) * options.Scale,
+				AccelerationPerFrame = new(Main.rand.NextFloat(-0.01f, 0.01f), -0.02f)
+			});
     }
 }
