+++ Common/DebuffOverhaul/Buffs/Venom.cs
@@ -1,14 +1,13 @@
 ﻿using SpiritReforged.Common.Particle;
-using SpiritReforged.DebuffOverhaul.Common;
-using SpiritReforged.DebuffOverhaul.Content.Particles;
+using SpiritReforged.Content.Particles.Basic;
 
-namespace SpiritReforged.DebuffOverhaul.Content.Buffs;
+namespace SpiritReforged.Common.DebuffOverhaul.Buffs;
 
-public class Poisoned : DoTExtension
+public class Venom : DoTExtension
 {
-    public override Settings LocalSettings => new(0.1f, 500);
+    public override Settings LocalSettings => new(0.5f, 1000);
 
-    public override void Load() => Handler.Register(this, BuffID.Poisoned);
+    public override void Load() => Handler.Register(this, BuffID.Venom);
 
     public override void PostDrawHealthBar(SpriteBatch spriteBatch, HealthBarHook.Options options)
     {
@@ -17,7 +16,7 @@ public class Poisoned : DoTExtension
         float fadeout = MathHelper.Min(BuffTime / 30f, 1);
         float lightness = options.Lightness * 2;
         Rectangle bounds = new(0, 0, (int)(front.Width * progress), front.Height);
-        Color color = new(0.6f, 1f, 0.2f);
+        Color color = Color.Purple;
 
         HealthBarHook.DrawSimpleBar(spriteBatch, front, options.Position, bounds, options.Scale, color * fadeout * lightness);
 
@@ -27,14 +26,12 @@ public class Poisoned : DoTExtension
 
         spriteBatch.Draw(bubble, endPosition, source, color * lightness, 0, source.Size() / 2, options.Scale, default, 0);
 
-        if ((int)Main.timeForVisualEffects % 18 == 0 && fadeout == 1)
-		{
-			TerrariaParticles.OverHealthBars.Add(new BubbleParticle(40, color * lightness, NPC)
-			{
-				LocalPosition = endPosition + Main.screenPosition - NPC.Center,
-				Scale = new Vector2(0.8f) * options.Scale,
-				AccelerationPerFrame = new(Main.rand.NextFloat(-0.01f, 0.01f), -0.02f)
-			});
-		}
+        if ((int)Main.timeForVisualEffects % 12 == 0 && fadeout == 1)
+            TerrariaParticles.OverHealthBars.Add(new BubbleParticle(30, color * lightness, NPC)
+            {
+                LocalPosition = Vector2.Lerp(options.Position + new Vector2(0, front.Height / 2), endPosition, Main.rand.NextFloat()) + Main.screenPosition - NPC.Center,
+                Scale = new Vector2(0.8f) * options.Scale,
+                AccelerationPerFrame = new(Main.rand.NextFloat(-0.01f, 0.01f), -0.02f)
+            });
     }
 }
