+++ Common/DebuffOverhaul/CustomDoT.cs
@@ -1,4 +1,4 @@
-﻿namespace SpiritReforged.DebuffOverhaul.Common;
+﻿namespace SpiritReforged.Common.DebuffOverhaul;
 
 /// <summary> A damage-over-time buff extension that can be registered through mod call. </summary>
 public class CustomDoT(float scalability, int damageLimit, Action<SpriteBatch, NPC, Color, Vector2, float, float> onPostDraw = null) : DoTExtension
