+++ Common/DebuffOverhaul/Textures/SmallFlame.png
